some code moved to Trait to clean up test class a bit
stan & cs
